backend/accounts/views.py
from django.db import models
from django.core import send_mail
import random

from django. contrib.auth.models import User
from rest_framework.authtoken.models import Token
from rest_framework import status
from rest_framework.response import Response
from rest_framework.permissions import AllowAny, IsAuthenticated, IsAdminUser
from rest_framework_simplejwt.tokens import TokenObtainPairView
import datetime 


