backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.10,<4.0
django-allauth>=0.44,<1.0
django-rest-auth>=0.9.5,<1.0
djangorestframework-simplejwt>=5.0,<6.0
django-phonenumber-field